+++ GWorkspace/Finder/SearchResults/SearchResults.m
@@ -730,14 +730,14 @@ static NSString *lsfname = @"LiveSearch.lsf";
 //
 // NSTableDataSource protocol
 //
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
   return [foundObjects count];
 }
 
 - (id)tableView:(NSTableView *)aTableView
           objectValueForTableColumn:(NSTableColumn *)aTableColumn
-                                row:(int)rowIndex
+                                row:(NSInteger)rowIndex
 {
   FSNode *node = [foundObjects objectAtIndex: rowIndex];
   
@@ -806,7 +806,7 @@ static NSString *lsfname = @"LiveSearch.lsf";
 - (void)tableView:(NSTableView *)aTableView 
   willDisplayCell:(id)aCell 
    forTableColumn:(NSTableColumn *)aTableColumn 
-              row:(int)rowIndex
+              row:(NSInteger)rowIndex
 {
   if (aTableColumn == nameColumn) {
     FSNTextCell *cell = (FSNTextCell *)[nameColumn dataCell];
